Workflow.md: name drop `git request-pull`
This is the builtin way to make pull requests via email.  Mentioning
it may lead exisiting `request-pull` users to the `hub` documentation
(which is roughly similar).  This is also something of a note to
myself, since I sometimes forget that `request-pull` exists.
`request-pull` has been in Git since quite early on:

  commit ab421d2c7886341c246544bc8dea43c66a44d909
  Author: Ryan Anderson <ryan@michonline.com>
  Date:   Tue Jul 26 03:30:36 2005 -0400

    [PATCH] Add git-request-pull-script, a short script that generates a summary of pending changes

    A short message requesting a pull from the repository is also included.

    Signed-off-by: Ryan Anderson <ryan@michonline.com>
    Signed-off-by: Junio C Hamano <junkio@cox.net>