_layouts/mapcal.html: Add FullCalendar-based calendar
FullCalendar is dual licensed MIT/GPLv2.  SWC uses the MIT license for
code, so including the FullCalendar code is allowed.  FullCalendar
itself uses JQuery, which is also MIT licensed.

Start dates are extracted from the `_posts/YYYY-MM-DD-...` filenames.
Day spans (for multi-day boot camps) come from a new `days` value in
the YAML front matter.  It should be possible to parse this out of the
less structured `dates` value, but that seemed like more work than it
was worth.

Incidental changes:
* Remove JQuery-latest import from footer-javascript.html, because it
  conflicted with the now-local jquery-1.8.1.
* Make the map a bit wider, because there's only so far you can
  scrunch FullCalendar before things start going wonky, and it's nice
  to have both the map and the calendar be the same width.

I used FullCalendar v1.5.4, which is the current release.