boot-camps.ics: Add template for Jekyll-generated iCalendar output
Create an iCalendar (RFC 5545) with event info for each boot camp.
This requires the following new YAML front matter fields:

  timestamp: the date on which the event was last modified (for
    DTSTAMP)
  end: the date when the event ends (for DTEND)

It also takes advantage of the existing `time`, which it uses for
DTSTART.  Both of these dates should be in ISO 8601 format
(YYYY-MM-DDThh:mm:ssTZD), but any timezone data given in TZD is
ignored, and the time is interpreted as UTC.  If Liquid's `date`
filter could convert input timestamps to UTC, we would have more
flexibility here, but it can't, and RFC 5545 requires UTC times for
some fields (e.g. DTSTAMP, see section 3.8.7.2).  Because of the
timezone issues, you should probably just stick to YYYY-MM-DD format,
and leave the time unspecified.

There is a good deal of overlap with the existing YAML front matter
`dates` and `days`.  In principle, `days` and `dates` could be
calculated from `time` and `end`, but Liquid doesn't expose enough
date-parsing logic for this to be feasible (as far as I can tell).

Here's the description of timestamp/DTSTAMP from RFC 5545:

   3.8.7.2. Date-Time Stamp
   Property Name:  DTSTAMP
   Purpose: ...In the case of an iCalendar object that doesn't specify
      a "METHOD" property, this property specifies the date and time
      that the information associated with the calendar component was
      last revised in the calendar store.

In principle, this could also be generated from the VCS log of the
post itself, but again we are limited to what we can accomplish in
pure Liquid.