CONTRIBUTING.md: Link to COPYING for explicit license details
The DCO refers to "the open source license(s) involved".  Give people
reading CONTRIBUTING.md a clear link to what that entails.  As implied
by the DCO, the license should also be listed in the file itself
("under the open source license indicated in the file"), but it
doesn't hurt to be explicit here.

Signed-off-by: W. Trevor King <wking@tremily.us>