modular/shell: Start reorganizing the shell material
Mostly split the examples out into their own directories with READMEs,
but there was a lot of editing to partially convert the long-form
notes into a condensed instructor.md targeted at instructors.  I've
gotten to the pipes section and run out of steam.  Erik may be taking
over the rest of the rewrite...