version-control/instructor.md: Add a 'Tool choices' section
Software Carpentry is about how to develop software, which is largely
independent of the tooling for a particular task.  This branch only
contains the generic motivation and introduction for version control.
That's nice, but a particular Software Carpentry course or boot camp
*is* going to want to teach a specific tool, and we want instructors
to easily be able to find and integrate notes for the tool they
choose.

This setup doesn't assume that there is a single canonical set of
notes for a given tool.  If you don't like the branches I've listed,
write up your own.  If you tell me about it, I'll add a link to your
branch as well, and we can try and work up some short descriptions to
help new instructors differentiate between the two approaches.

The choices are in the format used by `git request-pull`, so you can
run:

  $ git merge $CHOICE

to merge the branch.  For example:

  $ git merge git://github.com/wking/swc-boot-camps.git version-control-git

without bothering to setup remotes, etc.